Simplify connect and initial response read timeouts
When we connect to an SMTP server, we wait for two events: the initial TCP
connection completing (from the client's point of view, this is the
arrival of the SYN/ACK from the remote server) and receiving the initial
SMTP message (e.g. "220 OK").

Users of this library however, aren't able to do much with this
distinction. There's nothing productive to do until the initial response
is received, so that is the only timeout that really matters.

Futhermore, there was a bug in the previous implementation: if you set a
lower initialResponseReadTimeout than connectionTimeout, the client would
always see a failure even if the TCP connection was established.

Instead, we simplify the API by adding initialResponseTimeout.
This applies to the time from the initial SYN sent from the client to
the arrival of the initial SMTP response. This is what users of the API
likely want, and makes it impossible to misconfigure these timeouts.